fix: refactor step 5 so it's easier to work with
