+++ frontend/src/components/InfoSection/styles.scss
@@ -5,6 +5,7 @@
 
   input {
     cursor: default;
+    border-bottom: none;
   }
 
   .#{vars.$bcgov-prefix}--form-item {
